refactor(activities): the attachment write path gets its own file
`attachment.go` reached 506 lines, past the 500-line cap, when the upload grew a
streaming digest. The split is the seam that was already there: what one upload
carries, how its bytes are fingerprinted and measured, and the transaction that
lands the object and its row — separate from the read side, which answers a
different question, about scope over rows that already exist.
